refactor(dailyPokemon) - move dailyPokemon from business to utils
right now we have a bucket that's called utils for all stuff we don't know where to put yet.
